refactor/add_in_trade_fill_class
