+++ hummingbot/model/trade_fill.py
@@ -1,8 +1,14 @@
 #!/usr/bin/env python
-import numpy
+import numpy, pandas as pd
 from typing import (
     Any,
     Dict,
+List,
+)
+from hummingbot.core.event.events import (
+    TradeType,
+    TradeFee,
+    OrderType,
 )
 from sqlalchemy import (
     Column,
@@ -15,6 +21,7 @@ from sqlalchemy import (
     JSON
 )
 from sqlalchemy.orm import relationship
+from datetime import datetime
 
 from . import HummingbotBase
 
@@ -54,6 +61,39 @@ class TradeFill(HummingbotBase):
             f"quote_asset='{self.quote_asset}', timestamp={self.timestamp}, order_id='{self.order_id}', " \
             f"trade_type='{self.trade_type}', order_type='{self.order_type}', price={self.price}, " \
             f"amount={self.amount}, trade_fee={self.trade_fee}, exchange_trade_id={self.exchange_trade_id})"
+
+    @classmethod
+    def to_pandas(cls, trades: List):
+        columns: List[str] = ["symbol",
+                              "price",
+                              "amount",
+                              "order_type",
+                              "trade_type",
+                              "market",
+                              "timestamp",
+                              "fee_percent",
+                              "flat_fee / gas"]
+        data = []
+        for trade in trades:
+            if len(trade.trade_fee['flat_fees']) == 0:
+                flat_fee_str = "None"
+            else:
+                fee_strs = [f"{fee_tuple[0]} {fee_tuple[1]}" for fee_tuple in trade.trade_fee.flat_fees]
+                flat_fee_str = ",".join(fee_strs)
+
+            data.append([
+                trade.symbol,
+                trade.price,
+                trade.amount,
+                "market" if trade.order_type is OrderType.MARKET else "limit",
+                "buy" if trade.trade_type is TradeType.BUY else "sell",
+                trade.market,
+                datetime.fromtimestamp(int(trade.timestamp / 1e3)).strftime("%Y-%m-%d %H:%M:%S"),
+                trade.trade_fee['percent'],
+                flat_fee_str,
+            ])
+
+        return pd.DataFrame(data=data, columns=columns)
 
     @staticmethod
     def to_bounty_api_json(trade_fill: "TradeFill") -> Dict[str, Any]:
